requirements.txt
pytorch
pytorch-ignite
numpy
matplotlib
random
pickle
collections
nltk
gdown
